requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==4.0.1
django-cache-url==3.2.3
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.13.1
environs==8.0.0
gunicorn==20.1.0
marshmallow==3.14.1
mysqlclient==2.1.0
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0